Decide if sorter and visualizer are used
sorter and visualizer now supports function is_enabled that is called by
files_tracker. If the function returns true, the sorter/visualizer is
added to the dict. Otherwise, it is ignored.

That way, one can quickly add or remove options without changing file
names

The base classes sorter and visualizer are not enabled